se agrega condicional

+++ Jenkinsfile
@@ -11,15 +11,20 @@ pipeline{
    stages {
 
         stage('detener/limpiar'){
+	    
             steps{
-                sh '''
-                #sudo service nginx restart 
-                #sudo docker compose down
-                sudo docker stop ${name_container}
-                sudo docker rm ${name_container}
-                sudo docker system prune -f
-                sudo docker images purge
-                '''
+		      script {
+		          def dockerRuning = sh(script: "docker inspect ${name_container}| grep "Running"", returnStdout: true) == 0
+		          if(dockerRuning){
+			        sh '''
+            			sudo docker stop ${name_container}
+                        sudo docker rm ${name_container}
+                        sudo docker system prune -f
+                        sudo docker images purge
+            		'''
+		            }
+
+			    }
             }
         }
